Acabado caso de uso OrganizarPartido

+++ controller/OrganizeMatchController.php
@@ -12,8 +12,6 @@ require_once(__DIR__."/../model/ParticipantsMatchMapper.php");
 require_once(__DIR__."/../model/Reservation.php");
 require_once(__DIR__."/../model/ReservationMapper.php");
 
-require_once(__DIR__."/../model/ReservationOrganizeMatchMapper.php");
-
 require_once(__DIR__."/../controller/BaseController.php");
 
 /**
@@ -21,27 +19,35 @@ require_once(__DIR__."/../controller/BaseController.php");
 *
 * Controller to organize a match
 *
-* @author lipido <lipido@gmail.com>
 */
 class OrganizeMatchController extends BaseController {
 
   /**
   * Reference to the OrganizeMatchMapper to interact
   * with the database
-  *
   * @var OrganizeMatchMapper
   */
   private $organizeMatchMapper;
+
+  /**
+  * Reference to the ParticipantsMatchMapper to interact
+  * with the database
+  * @var ParticipantsMatchMapper
+  */
   private $participantsMatchMapper;
+
+  /**
+  * Reference to the ReservationMapper to interact
+  * with the database
+  * @var ReservationMapper
+  */
   private $reservationMapper;
-  private $reservationOrganizeMatchMapper;
 
   public function __construct() {
     parent::__construct();
     $this->organizeMatchMapper = new OrganizeMatchMapper();
     $this->participantsMatchMapper =  new ParticipantsMatchMapper();
     $this->reservationMapper = new ReservationMapper();
-    $this->reservationOrganizeMatchMapper = new ReservationOrganizeMatchMapper();
   }
 
 
@@ -53,60 +59,51 @@ class OrganizeMatchController extends BaseController {
       $organizeMatch = new OrganizeMatch();
 
       if (isset($_POST["dateOrganizeMatch"])) { // reaching via HTTP Post...
-
-        // populate the Post object with data form the form
-        $organizeMatch->setFecha(date("Y,m,d", strtotime($_POST["dateOrganizeMatch"])));
+        $organizeMatch->setFecha(date("Y-m-d", strtotime($_POST["dateOrganizeMatch"])));
         $organizeMatch->setHora($_POST["timeOrganizeMatch"]);
 
-        try {
-          // validate OrganizeMatch object
-          $organizeMatch->checkIsValidForCreate(); // if it fails, ValidationException
 
-          // save the OrganizeMatch object into the database
+
+        $numRes = $this->reservationMapper->getNumReservations($organizeMatch->getFecha(), $organizeMatch->getHora());
+        if( $numRes == 5 ){
+          throw new Exception(" Alredy five matches for this date");
+        }
+
+        try {
+          $organizeMatch->checkIsValidForCreate();
+
           $this->organizeMatchMapper->save($organizeMatch);
 
-          // POST-REDIRECT-GET
-          // Everything OK, we will redirect the user to the list of posts
-          // We want to see a message after redirection, so we establish
-          // a "flash" message (which is simply a Session variable) to be
-          // get in the view after redirection.
           $this->view->setFlash(sprintf(i18n("Match successfully organize.")));
 
-          // perform the redirection. More or less:
-          // header("Location: index.php?controller=users&action=index")
-          // die();
           $this->view->redirect("users", "index");
 
         }catch(ValidationException $ex) {
-          // Get the errors array inside the exepction...
           $errors = $ex->getErrors();
-          // And put it to the view as "errors" variable
           $this->view->setVariable("errors", $errors);
         }
       }
 
-      // Put the OrganizeMatch object visible to the view
       $this->view->setVariable("organizeMatch", $organizeMatch);
 
-      // render the view (/view/organizeMatch/add.php)
       $this->view->render("organizeMatch", "add");
 
     }
+
 
     public function viewAll() {
       if (!isset($this->currentUser)) {
         throw new Exception("Not in session. You must be logged");
       }
 
+      $this->validateAllOrganizeMatches();
+
       $organizedMatches = $this->organizeMatchMapper->findAll();
 
-      // Put the OrganizeMatch object visible to the view
       $this->view->setVariable("organizedMatches", $organizedMatches);
 
-      // render the view (/view/organizeMatch/add.php)
       $this->view->render("organizeMatch", "view");
     }
-
 
 
     public function join(){
@@ -134,43 +131,38 @@ class OrganizeMatchController extends BaseController {
 
         try{
 
-          // validate OrganizeMatch object
-          $participant->checkIsValidForCreate(); // if it fails, ValidationException
+          $participant->checkIsValidForCreate();
 
-          // save the OrganizeMatch object into the database
-          $this->participantsMatchMapper->save($participant);
+          $this->participantsMatchMapper->save($participant); // Error 4 insercion
 
-          // POST-REDIRECT-GET
-          // Everything OK, we will redirect the user to the list of posts
-          // We want to see a message after redirection, so we establish
-          // a "flash" message (which is simply a Session variable) to be
-          // get in the view after redirection.
           $this->view->setFlash(sprintf(i18n("Match successfully join.")));
+
+
 
           $players = $this->participantsMatchMapper->count($idOrganizeMatch);
 
           if( $players == 4 ){
             $organizeMatch = $this->organizeMatchMapper->find($idOrganizeMatch);
             if($organizeMatch != null){
-              $freePista = $this->reservationMapper->getNumReservations($organizeMatch->getFecha(), $organizeMatch->getHora());
-              $reservation = new Reservation(null, "admin", $organizeMatch->getFecha(), $organizeMatch->getHora(), $freePista+1);
-              $this->reservationMapper->makeReservation($reservation);
-
-              $this->reservationOrganizeMatchMapper->save( $this->reservationMapper->getReservationId("admin", $organizeMatch->getFecha(), $organizeMatch->getHora(), $freePista+1),
-                                                                  $idOrganizeMatch);
-              $this->organizeMatchMapper->delete($idOrganizeMatch);
+              $numRes = $this->reservationMapper->getNumReservations($organizeMatch->getFecha(), $organizeMatch->getHora());
+              if( $numRes != 5 ){
+                $reservation = new Reservation(null, "admin", $organizeMatch->getFecha(), $organizeMatch->getHora());
+                $this->reservationMapper->makeReservation($reservation);
+                $this->organizeMatchMapper->delete($idOrganizeMatch);
+              }
+              else{
+                throw new Exception(" Alredy five matches for this date");
+              }
+            }
+            else {
+              throw new Exception(" Organize match not found");
             }
           }
 
-          // perform the redirection. More or less:
-          // header("Location: index.php?controller=users&action=index")
-          // die();
           $this->view->redirect("users", "index");
 
         }catch(ValidationException $ex) {
-          // Get the errors array inside the exepction...
           $errors = $ex->getErrors();
-          // And put it to the view as "errors" variable
           $this->view->setVariable("errors", $errors);
         }
       }
@@ -178,13 +170,10 @@ class OrganizeMatchController extends BaseController {
       $joinMatch = $this->organizeMatchMapper->findMatchWithParticipants($idOrganizeMatch);
       $play = $this->participantsMatchMapper->play($idOrganizeMatch, $_SESSION["currentuser"]);
 
-      // Put $play true if the player alredy play in this game, and false if the play not play
       $this->view->setVariable("play", $play );
 
-      // Put the OrganizeMatch object visible to the view
       $this->view->setVariable("joinMatch", $joinMatch );
 
-      // render the view (/view/organizeMatch/add.php)
       $this->view->render("organizeMatch", "join");
 
     }
@@ -206,35 +195,22 @@ class OrganizeMatchController extends BaseController {
         }
 
         try{
-          // save the OrganizeMatch object into the database
           $this->organizeMatchMapper->delete($_POST["idOrganizeMatch"]);
 
-          // POST-REDIRECT-GET
-          // Everything OK, we will redirect the user to the list of posts
-          // We want to see a message after redirection, so we establish
-          // a "flash" message (which is simply a Session variable) to be
-          // get in the view after redirection.
           $this->view->setFlash(sprintf(i18n("Match successfully delete.")));
 
-          // perform the redirection. More or less:
-          // header("Location: index.php?controller=users&action=index")
-          // die();
           $this->view->redirect("users", "index");
 
         }catch(ValidationException $ex) {
-          // Get the errors array inside the exepction...
           $errors = $ex->getErrors();
-          // And put it to the view as "errors" variable
           $this->view->setVariable("errors", $errors);
         }
       }
 
       $deleteMatch = $this->organizeMatchMapper->findMatchWithParticipants($idOrganizeMatch);
 
-      // Put the OrganizeMatch object visible to the view
       $this->view->setVariable("deleteMatch", $deleteMatch );
 
-      // render the view (/view/organizeMatch/add.php)
       $this->view->render("organizeMatch", "delete");
 
     }
@@ -252,25 +228,14 @@ class OrganizeMatchController extends BaseController {
 
         try{
 
-          // save the OrganizeMatch object into the database
           $this->participantsMatchMapper->cancel($_POST["idOrganizeMatch"], $_SESSION["currentuser"]);
 
-          // POST-REDIRECT-GET
-          // Everything OK, we will redirect the user to the list of posts
-          // We want to see a message after redirection, so we establish
-          // a "flash" message (which is simply a Session variable) to be
-          // get in the view after redirection.
           $this->view->setFlash(sprintf(i18n("Cancel successfully.")));
 
-          // perform the redirection. More or less:
-          // header("Location: index.php?controller=users&action=index")
-          // die();
           $this->view->redirect("users", "index");
 
         }catch(ValidationException $ex) {
-          // Get the errors array inside the exepction...
           $errors = $ex->getErrors();
-          // And put it to the view as "errors" variable
           $this->view->setVariable("errors", $errors);
         }
       }
@@ -278,4 +243,15 @@ class OrganizeMatchController extends BaseController {
         throw new Exception("ID is mandatory");
       }
     }
+
+    public function validateAllOrganizeMatches(){
+      $organizedMatches = $this->organizeMatchMapper->findAll();
+      foreach ($organizedMatches as $organizeMatch) {
+        $numRes = $this->reservationMapper->getNumReservations($organizeMatch->getFecha(), $organizeMatch->getHora());
+        if( $numRes == 5 ){
+          $this->organizeMatchMapper->delete($organizeMatch->getIdOrganizarPartido());
+        }
+      }
+    }
+
   }
